+++ packages/client/src/pages/forum/forumPick/ForumPick.tsx
@@ -1,38 +1,17 @@
 import React, {useEffect, useState} from 'react';
-import styles from './scss/forumPick.module.scss';
+import styles from './forumPick.module.scss';
 import { Link } from 'react-router-dom';
-import { MOCK_ARRAY } from './mockList'
+import { MOCK_ARRAY } from './mockList';
+import Modal from "./modules/modal/modal";
 
 interface IMockForumList {ID: string, TITLE: string}
 
 const ForumPick = () => {
-
-  const [textState, setTextState] = useState<string>("");
-  const [isHidden, setIsHidden] = useState<boolean>(true);
+  const [isModalHidden, setIsModalHidden] = useState<boolean>(true);
   const [forumList, setForumList] = useState<IMockForumList[]>([]);
 
   function showModalEvent() {
-    setIsHidden(false)
-  }
-  
-  function submitEvent(e: React.SyntheticEvent) {
-    e.preventDefault();
-    if (!textState) {
-      alert("Enter the title")
-      return
-    }
-    alert(`Forum with the title "${textState}" was created`)
-    setIsHidden(true)
-    setTextState("")
-  }
-  
-  function cancelEvent() {
-    setIsHidden(true)
-    setTextState("")
-  }
-
-  function changeEvent(e: React.ChangeEvent<HTMLTextAreaElement>) {
-    setTextState(e.target.value)
+    setIsModalHidden(false)
   }
 
   useEffect(() => {
@@ -42,14 +21,7 @@ const ForumPick = () => {
 
   return (
     <div className={styles.forumListContainer}>
-      <div className={styles.modalWrapper} style={{display: isHidden ? "none" : "flex"}}>
-        <form className={styles.modal}>
-          <h2 className={styles.title}>Enter forum's title</h2>
-          <textarea value={textState} onChange={changeEvent} className={styles.textBlock} name="forumName" placeholder="Forum's title"></textarea>
-          <button type="submit" className={styles.add} onClick={submitEvent}>Submit</button>
-          <button type="button" className={styles.cancel} onClick={cancelEvent}>Cancel</button>
-        </form>
-      </div>
+      {!isModalHidden && <Modal setIsModalHidden={setIsModalHidden}/>}
       <div className={styles.forumList}>
         <div onClick={showModalEvent} className={styles.link}>
           <p className={styles.addLabel}>
