+++ packages/client/src/pages/forum/forumPick/forumPick.module.scss
@@ -1,5 +1,4 @@
-@import '../../globals.scss';
-@import './modal.scss';
+@import '../globals.scss';
 
 body {
 	margin: 0;
