+++ packages/client/src/pages/forum/forumPick/mockList.ts
@@ -1,14 +1,14 @@
 export const MOCK_ARRAY = [
-    {
-      TITLE: "Forum title",
-      ID: "gdf4"
-    },
-    {
-      TITLE: "Forum title",
-      ID: "hf2n"
-    },
-    {
-      TITLE: "Forum title",
-      ID: "g8lolk"
-    }
+  {
+    TITLE: "Forum title",
+    ID: "gdf4"
+  },
+  {
+    TITLE: "Forum title",
+    ID: "hf2n"
+  },
+  {
+    TITLE: "Forum title",
+    ID: "g8lolk"
+  }
 ]
