+++ packages/client/src/pages/forum/forumPick/modules/modal/modal.module.scss
@@ -1,7 +1,7 @@
-@import '../../globals.scss';
+@import '../../../globals.scss';
 
 .modalWrapper {
-	display: none;
+	display: flex;
 	position: fixed;
 	width: 100%;
 	height: 100%;
